# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "You can visit out RiderTrack page at [https://rider-track-dev.herokuapp.com]" 
  version: "1.0.0"
  title: "Ridertrack Swagger"
host: "rider-track-dev.herokuapp.com"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /auth/login:
    post:
      tags:
      - "authentication"
      summary: "Sign in an user"
      description: "It checks the given email and password in the db. If matches it creates a jwt and return it."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Object with parameters(email and password) used to sign in an user"
        required: true
        schema:
          description: ""
      responses:
        200:
          description: ""
          headers:
            userId:
              type: "string"
              description: "Id of signed user"
            role:
              type: "string"
              description: "role of signed user"
            jwt-token:
              type: "string"
              description: "jwt token for this user"
            expiresIn:
              type: "number"
              description: "time in which token expires in this session"
        400:
          description: "Error message"
  /auth/login/facebook:
    get:
      tags:
      - "authentication"
      summary: "Sign in as an facebook user"
      description: "Sign in an user with facebook authentication"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: ""
          headers:
            userId:
              type: "string"
              description: "Id of signed user"
            role:
              type: "string"
              description: "role of signed user"
            jwt-token:
              type: "string"
              description: "jwt token for this user"
            expiresIn:
              type: "number"
              description: "User is not found or password is incorrect"
        400:
          description: "Error message"
  /auth/login/google:
    get:
      tags:
      - "authentication"
      summary: "Sign in as an google user"
      description: "Sign in an user with google authentication"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: ""
          headers:
            userId:
              type: "string"
              description: "Id of signed user"
            role:
              type: "string"
              description: "role of signed user"
            jwt-token:
              type: "string"
              description: "jwt token for this user"
            expiresIn:
              type: "number"
              description: "time in which token expires in this session"
        400:
          description: "Error message"
...